+++ stories/index.tsx
@@ -1,38 +1,46 @@
-import * as React from 'react'
-import { storiesOf } from '@storybook/react'
+import * as React from 'react';
+import { storiesOf } from '@storybook/react';
 
-import BreakpointObserver from 'breakpoint-observer'
+import BreakpointObserver from 'breakpoint-observer';
 
-storiesOf('breakpoint-observer', module).add('Current Width', () => (
+storiesOf('breakpoint-observer', module).add('Current Breakpoint', () => (
   <React.Fragment>
-    <h2>Minimal config</h2>
-    <p>Without any props <code>{`<BreakpointObserver />`}</code> returns current the <code>window.innerWidth</code></p>
+    <h2>Observe Current Breakpoint</h2>
+    <p>
+      Supplying a breakpoint object to <code>{`<BreakpointObserver />`}</code>{' '}
+      will return the current breakpoint.
+    </p>
     <pre>
-    <code>
-{`<BreakpointObserver>
-  {({ width }) => <p>Current width is {width} pixels!</p>}
+      <code>
+        {`<BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }}>
+  {breakpoint => <p>The current breakpoint is {breakpoint}!</p>}
 </BreakpointObserver>`}
-    </code>
-    </pre>
-    <BreakpointObserver>
-      {({ width }) => <p>Current width is {width} pixels!</p>}
-    </BreakpointObserver>
-  </React.Fragment>
-))
-
-storiesOf('breakpoint-observer', module).add('Breakpoints', () => (
-  <React.Fragment>
-    <h2>Breakpoints</h2>
-    <p>Supplying a breakpoints object to <code>{`<BreakpointObserver />`}</code> will return the current breakpoint.</p>
-    <pre>
-    <code>
-{`<BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }}>
-  {({ breakpoint }) => <p>Current breakpoint is {breakpoint}!</p>}
-</BreakpointObserver>`}
-    </code>
+      </code>
     </pre>
     <BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }}>
-      {({ breakpoint }) => <p>Current breakpoint is {breakpoint}!</p>}
+      {breakpoint => <p>Current breakpoint is {breakpoint}!</p>}
     </BreakpointObserver>
   </React.Fragment>
-))
+));
+
+storiesOf('breakpoint-observer', module).add('Default Breakpoint', () => (
+  <React.Fragment>
+    <h2>Render Server-Side With a Default Breakpoint</h2>
+    <p>
+      In environments without a window, a default breakpoint can be supplied.
+    </p>
+    <pre>
+      <code>
+        {`<BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }} defaultBreakpoint={'tablet'}>
+  {breakpoint => <p>The current breakpoint is {breakpoint}!</p>}
+</BreakpointObserver>`}
+      </code>
+    </pre>
+    <BreakpointObserver
+      breakpoints={{ mobile: 0, tablet: 768 }}
+      defaultBreakpoint={'tablet'}
+    >
+      {breakpoint => <p>Current breakpoint is {breakpoint}!</p>}
+    </BreakpointObserver>
+  </React.Fragment>
+));