+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2018-03-23
+### Removed
+- Remove `window.innerWidth` support to focus on breakpoints only
+### Added
+- Add support for a default breakpoint for SSR
+
 ## [0.2.1] - 2018-03-22
 ### Changed
 - Better Typescript definitions